move file types.ts and create RootState
