+++ src/app/redux/appReducer.ts
@@ -3,7 +3,9 @@ import { combineReducers } from 'redux';
 import cartReducer from '../pages/cart/cart.reducer';
 
 const rootReducer = combineReducers({
-  cartReducer: cartReducer
+  cart: cartReducer
 });
 
 export default rootReducer;
+
+export type RootState = ReturnType<typeof rootReducer>;
