+++ src/app/shared/components/layout/Header.tsx
@@ -1,11 +1,12 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import { useSelector } from 'react-redux';
+import { RootState } from '../../../redux/appReducer';
 import Icon from '../../../../assets/icon';
 import { ICart } from './../../interfaces/cart';
 
 export const Header = () => {
-  const { cart } = useSelector((state: any) => state.cartReducer);
+  const { cart } = useSelector((state: RootState) => state.cart);
 
   const cartNumber = cart.reduce((acc: number, item: ICart) => acc + item.quantity, 0);
   return (
